Correct `Error` implementation of `std::error::Error::source()`
I'm so excited about figuring this out that I'm going to write a
multi-line commit message. Hold onto your hats, folks, this is gonna get
nerdy.

In past years Rust implemented the Error trait roughly as follows:

```rust
trait std::error::Error {
    fn cause(&self) -> Option<impl Error>;
}
```

(Or something like that. I'm a bit rusty with Rust.)
(OH LOOK A SUBTLE PUN AND IT'S RIGHT THERE HOO BOY WE GOTTA MAKE THE JOKE)

I generally implemented this with a concrete custom type, so life was
simple:

```rust
struct ErrmagErrd {
    inner: std::io::Error,
}

impl ErrmagErrd {
    pub fn new(inner: std::io::Error) -> Self {
        ErrmagErrd { inner }
    }
}

impl std::error::Error for ErrmagErrd {
    fn cause(&self) -> impl Error {
        // HURR HURR RUST IS AMAZING
        // I DON'T KNOW WHAT I'M DOING
        // BUT I'M HAVING SO MUCH FUN
        Ok(&self.inner)
    }
}
```

So. Less than elegant. But it worked and it was simple.

PART 2.

So time past, I got a bit better at Rust, and then life happened and I
forgot a bunch.

In the mean time the `std::error::Error` trait has changed:

```rust
trait std::error::Error {
    fn source(&self) -> Option<&(dyn Error + 'static)> { None }
}
```

I had heard of dyn but had not used it, so that's a bit of fun. Second,
uhhhh `'static`? What is that referring to? (Best of my ability, it may
be the trait itself or the definition of that trait.)

But I have no idea.

I also had switched the struct definition to make the source optional:

```rust
struct Error {
  source: Option<Box<&(dyn std::error::Error)>>,
}
```

Dyn makes sense, but I had not used it.

Sooooo.

But this codebase, my dear people, is meant to be as much a work of art
as is the Mandelbrot itself. Also I love Rust and I never get a chance
to write it, so this project is going to get way too much attention.

So long story short I've been trying to figure out how to get this
working in a clean fashion that I understand rather than rocking the
cargo culting.

I tried shit like so:

```rust
self.source.map(|e| e.as_ref())
```

Or other weird assemblages of syntax I didn't understand.

This definition for `std::io::Error` has the following:

```rust
    fn source(&self) -> Option<&(dyn error::Error + 'static)> {
        match self.repr {
            Repr::Os(..) => None,
            Repr::Simple(..) => None,
            Repr::Custom(ref c) => c.error.source(),
        }
```

This reduces to the following, which worked:

```
    fn source(&self) -> Option<&(dyn error::Error + 'static)> {
        match self {
            None => None, // Hacky - this feels like code smoke
            Some(ref e) => e.source()
        }
    }
```

It should not be necessary to use a `match` expression to unpack the
structure. In addition this was broken because I was burrowing two
`source()` calls down, so this is broken.

So.

Here's what we've got.

```
impl std::error::Error for Error {
    fn source(&self) -> Option<&(dyn std::error::Error + 'static)> {
        self.source.as_ref().map(|e| &**e)
    }
}
```

This is so neat I wanted to document it. Might as well a commit message.

Thank you for coming to my TED talk.